chackers/requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
gunicorn==22.0.0
pydantic_settings==2.1.0
pydantic[email]==2.5.2
fastapi-sessions==0.3.2
python-dotenv~=1.0.1